able to delete files

+++ Section 20 - File Upload & Download/controllers/admin.js
@@ -1,4 +1,7 @@
 const mongoose = require("mongoose");
+
+const fileHelper = require("../util/file");
+
 const { validationResult } = require("express-validator/check");
 
 const Product = require("../models/product");
@@ -172,8 +175,9 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDescription;
 
-      // if no new image was passed
+      // if new image was passed
       if (image) {
+        fileHelper.deleteFile(product.imageUrl); // deleting the file that is being replaced
         product.imageUrl = image.path;
       }
 
@@ -215,7 +219,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+  // for deleting a product
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found."));
+      }
+      fileHelper.deleteFile(product.imageUrl); // removes the associated file
+
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
